requirements.txt
streamlit==1.37.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
transformers==4.43.3
torch>=2.1.0
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2
python-docx==1.1.2
pypdf==4.3.1
tiktoken==0.7.0
openai==1.37.1
